39: Fixed nested devices syncing in DevicesPanel
The devices panel was not updating the state of the children's devices after a parent
state change. This was due to calling the wrong dependencies in the useEffect hook.

The logic that managed the PowerSwitches and Sliders in the devicesReducer was
also improved to correctly apply the state changes in synchronicity and to disallow
it when no sync was meant to happen.

Also, the bedroom icon was replaced with a color version.