39: Prepared DevicesPanel for dynamic data fetching
DevicesPanel was enabled to fetch and render either all the devices in
a house or the devices specific to a room. Made the panel's title
dynamic too.

Also, a new field was added to the Device component to show the device
type name, in preparation for the 2nd batch of User Stories.

Other improvements:
 - Updated the SmartPlug icon
 - Moved all .css files into the CSS folder, and moved it out to /src root
 - Removed the local imports of images and set their source to the
   public /img folder. The /img folder inside /src was deleted.