#48: Created new devices (blinds, thermostat, and security camera)
Blinds are fully functional, security camera needs the modal in which the video
feed will be played, and for the thermostat, displaying the own device's temp
value and the average of the temp sensors in the room is being programmed.

Also, included the new icons for the devices, and for the two new dashboard
option: automation and guests.