#90: Managed conflicts between same light intensity type effects
Originally, there was the problem that when an user chose more than one light
intensity effect, the UI wasnt tracking which devices where selected for that
effect, leaving the door open to conflicting configurations. Now, as the devices
are chosen to be controlled by a an effect, they are removed from the available
devices list on the right side of any other effects of the same type.